+++ Sources/ClashBar/App/Session/AppSessionTypes.swift
@@ -44,7 +44,6 @@ enum SystemProxyHelperActionState: Equatable {
     case idle
     case installing
     case reinstalling
-    case resigningReinstalling
 }
 
 enum SystemProxyHelperIssue: Equatable {
@@ -53,7 +52,6 @@ enum SystemProxyHelperIssue: Equatable {
     case registrationFailed
     case systemPolicyBlocked
     case signatureMismatch
-    case missingSigningIdentity
     case needsApproval
     case installLocationInvalid
     case helperMissing